Add workaround for "highest level player is missing in additional.."
If your guild is strong enough to completely destroy the enemy, not all fighters will fight. Since some backend changes in roughly mid 2023 playa added an extra header inside the fight for players who did not fought called 'fightadditionalplayers'. This header contains a list of all players who did not have one single fight, excluding for unknown reasons the player with the highest level. With this new addition of the checkFight command you can ignore this player for the check and don't need to remove him manually.